extract Signature into its own module
